requirements.txt
flet==0.21.2
requests==2.31.0
Pillow==10.1.0